backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
aiofiles
python-multipart
pytesseract
pillow
pdf2image
PyPDF2